Finish system utils MacOS porting

+++ gradle47.sh
@@ -1,11 +1,40 @@
 #!/bin/bash
 
-wget https://downloads.gradle.org/distributions/gradle-4.7-bin.zip
-sudo mkdir /opt/gradle
-sudo unzip -d /opt/gradle gradle-4.7-bin.zip
-rm gradle-4.7-bin.zip
+source platform
+source prettyecho
 
+case $PLATFORM in
+  Debian*)
+    if [ -d /opt/gradle/gradle-4.7 ]; then
+      log_info 'Found existing gradle-4.7 installation. Skipping...'
+      exit 0
+    fi
+
+    wget https://downloads.gradle.org/distributions/gradle-4.7-bin.zip
+    sudo mkdir /opt/gradle
+    sudo unzip -d /opt/gradle gradle-4.7-bin.zip
+    rm gradle-4.7-bin.zip
+    ;;
+
+  Darwin*)
+    brew install gradle
+    ;;
+
+  *)
+    log_error "Platform not supported"
+    exit 1
+    ;;
+
+esac
+
+# TODO: Fix MacOS GRADLE_HOME variable
+# .profile should remain platform independent
 if [[ -z $(grep 'GRADLE_HOME' ~/.profile) ]]; then
-  echo 'GRADLE_HOME=/opt/gradle/gradle-4.7/bin' >> ~/.profile
+  echo 'case $(uname -s) in' >> ~/.profile
+  echo '  Linux*) GRADLE_HOME=/opt/gradle/gradle-4.7/bin ;;' >> ~/.profile
+  echo '  Darwin*) GRADLE_HOME="" ;;' >> ~/.profile
+  echo 'esac' >> ~/.profile
   echo 'PATH=$PATH:$GRADLE_HOME' >> ~/.profile
 fi
+
+
